+++ src/scripts/blog/slug.js
@@ -9,15 +9,6 @@ import '../components/content-viewer/index'
 import '../components/subscription/index'
 import '../components/blog-pressroom-viewer/index'
 
-import hljs from 'highlight.js'
+import { setupHighlightJS } from '../utils/setup-hljs'
 
-const highlightLineNumbersLoad = async () => {
-  return await import('highlightjs-line-numbers.js')
-}
-
-;(async function () {
-  hljs.highlightAll()
-  window.hljs = hljs
-  await highlightLineNumbersLoad()
-  hljs.initLineNumbersOnLoad()
-})()
+setupHighlightJS()
